Effects.Concurrent.Async

Description

Provides the MonadAsync typeclass for async effects. We first present the class, then the async API implemented via the typeclass.

Since: 0.1

Effect

class Monad m => MonadAsync m where

Represents async effects. API largely follows unliftio's implementation of UnliftIO.Async.

Since: 0.1

Methods

withAsync :: HasCallStack => m a -> (Async a -> m b) -> m b

Lifted withAsync.

Since: 0.1

withAsyncBound :: HasCallStack => m a -> (Async a -> m b) -> m b

Lifted withAsyncBound.

Since: 0.1

withAsyncOn :: HasCallStack => Int -> m a -> (Async a -> m b) -> m b

Lifted withAsyncOn.

Since: 0.1

withAsyncWithUnmask

Arguments

:: HasCallStack

=> ((forall c. m c -> m c) -> m a)
.

-> (Async a -> m b)

-> m b

Lifted withAsyncWithUnmask.

Since: 0.1

withAsyncOnWithUnmask

Arguments

:: HasCallStack

=> Int
.

-> ((forall c. m c -> m c) -> m a)

-> (Async a -> m b)

-> m b

Lifted withAsyncOnWithUnmask.

Since: 0.1

wait :: HasCallStack => Async a -> m a

Lifted wait.

Since: 0.1

poll :: HasCallStack => Async a -> m (Maybe (Either SomeException a))

Lifted poll.

Since: 0.1

waitCatch :: HasCallStack => Async a -> m (Either SomeException a)

Lifted waitCatch.

Since: 0.1

cancel :: HasCallStack => Async a -> m ()

Lifted cancel.

Since: 0.1

uninterruptibleCancel :: HasCallStack => Async a -> m ()

Lifted uninterruptibleCancel.

Since: 0.1

cancelWith :: (Exception e, HasCallStack) => Async a -> e -> m ()

Lifted cancelWith.

Since: 0.1

race :: HasCallStack => m a -> m b -> m (Either a b)

Lifted race.

Since: 0.1

concurrently :: HasCallStack => m a -> m b -> m (a, b)

Lifted concurrently.

Since: 0.1

concurrently_ :: HasCallStack => m a -> m b -> m ()

Lifted concurrently_.

Since: 0.1

waitAny :: HasCallStack => [Async a] -> m (Async a, a)

Lifted waitAny.

Since: 0.1

waitAnyCatch

Arguments

:: HasCallStack

=> [Async a]
.

-> m (Async a, Either SomeException a)

Lifted waitAnyCatch.

Since: 0.1

waitAnyCancel :: HasCallStack => [Async a] -> m (Async a, a)

Lifted waitAnyCancel.

Since: 0.1

waitAnyCatchCancel

Arguments

:: HasCallStack

=> [Async a]
.

-> m (Async a, Either SomeException a)

Lifted waitAnyCatchCancel.

Since: 0.1

waitEither :: HasCallStack => Async a -> Async b -> m (Either a b)

Lifted waitEither.

Since: 0.1

waitEitherCatch

Arguments

:: HasCallStack

=> Async a
.

-> Async b

-> m (Either (Either SomeException a) (Either SomeException b))

Lifted waitEitherCatch.

Since: 0.1

waitEitherCancel :: HasCallStack => Async a -> Async b -> m (Either a b)

Lifted waitEitherCancel.

Since: 0.1

waitEitherCatchCancel

Arguments

:: HasCallStack

=> Async a
.

-> Async b

-> m (Either (Either SomeException a) (Either SomeException b))

Lifted waitEitherCatchCancel.

Since: 0.1

waitEither_ :: HasCallStack => Async a -> Async b -> m ()

Lifted waitEither_.

Since: 0.1

waitBoth :: HasCallStack => Async a -> Async b -> m (a, b)

Lifted waitBoth.

Since: 0.1

async :: HasCallStack => m a -> m (Async a)

Lifted async.

Since: 0.1

asyncBound :: HasCallStack => m a -> m (Async a)

Lifted asyncBound.

Since: 0.1

asyncOn :: HasCallStack => Int -> m a -> m (Async a)

Lifted asyncOn.

Since: 0.1

asyncWithUnmask :: HasCallStack => ((forall b. m b -> m b) -> m a) -> m (Async a)

Lifted asyncWithUnmask.

Since: 0.1

asyncOnWithUnmask :: HasCallStack => Int -> ((forall b. m b -> m b) -> m a) -> m (Async a)

Lifted asyncOnWithUnmask.

Since: 0.1

link :: HasCallStack => Async a -> m ()

Lifted link.

Since: 0.1

linkOnly :: HasCallStack => (SomeException -> Bool) -> Async a -> m ()

Lifted linkOnly.

Since: 0.1

link2 :: HasCallStack => Async a -> Async b -> m ()

Lifted link2.

Since: 0.1

link2Only :: HasCallStack => (SomeException -> Bool) -> Async a -> Async b -> m ()

Lifted link2Only.

Since: 0.1

pooledMapConcurrentlyN

Arguments

:: (HasCallStack, Traversable t)

=> Positive Int
.

-> (a -> m b)

-> t a

-> m (t b)

Lifted pooledMapConcurrentlyN.

Since: 0.1

pooledMapConcurrently

Arguments

:: (HasCallStack, Traversable t)

=> (a -> m b)
.

-> t a

-> m (t b)

Lifted pooledMapConcurrently.

Since: 0.1

pooledMapConcurrentlyN_

Arguments

:: (Foldable f, HasCallStack)

=> Positive Int
.

-> (a -> m b)

-> f a

-> m ()

Lifted pooledMapConcurrentlyN_.

Since: 0.1

pooledMapConcurrently_

Arguments

:: (Foldable f, HasCallStack)

=> (a -> m b)
.

-> f a

-> m ()

Lifted pooledMapConcurrently_.

Since: 0.1

Asynchronous Actions

data Async a

An asynchronous action spawned by async or withAsync. Asynchronous actions are executed in a separate thread, and operations are provided for waiting for asynchronous actions to complete and obtaining their results (see e.g. wait).

High-level API

Spawning with automatic cancelation

withAsync :: (MonadAsync m, HasCallStack) => m a -> (Async a -> m b) -> m b

Lifted withAsync.

Since: 0.1

withAsyncBound :: (MonadAsync m, HasCallStack) => m a -> (Async a -> m b) -> m b

Lifted withAsyncBound.

Since: 0.1

withAsyncOn :: (MonadAsync m, HasCallStack) => Int -> m a -> (Async a -> m b) -> m b

Lifted withAsyncOn.

Since: 0.1

withAsyncWithUnmask

Arguments

:: (MonadAsync m, HasCallStack)

=> ((forall c. m c -> m c) -> m a)
.

-> (Async a -> m b)

-> m b

Lifted withAsyncWithUnmask.

Since: 0.1

withAsyncOnWithUnmask

Arguments

:: (MonadAsync m, HasCallStack)

=> Int
.

-> ((forall c. m c -> m c) -> m a)

-> (Async a -> m b)

-> m b

Lifted withAsyncOnWithUnmask.

Since: 0.1

Querying Asyncs

wait :: (MonadAsync m, HasCallStack) => Async a -> m a

Lifted wait.

Since: 0.1

poll :: (MonadAsync m, HasCallStack) => Async a -> m (Maybe (Either SomeException a))

Lifted poll.

Since: 0.1

waitCatch :: (MonadAsync m, HasCallStack) => Async a -> m (Either SomeException a)

Lifted waitCatch.

Since: 0.1

asyncThreadId :: Async a -> ThreadId

Returns the ThreadId of the thread running the given Async.

cancel :: (MonadAsync m, HasCallStack) => Async a -> m ()

Lifted cancel.

Since: 0.1

uninterruptibleCancel :: (MonadAsync m, HasCallStack) => Async a -> m ()

Lifted uninterruptibleCancel.

Since: 0.1

cancelWith :: (MonadAsync m, Exception e, HasCallStack) => Async a -> e -> m ()

Lifted cancelWith.

Since: 0.1

data AsyncCancelled

The exception thrown by cancel to terminate a thread.

Constructors

AsyncCancelled

High-level utilities

race :: (MonadAsync m, HasCallStack) => m a -> m b -> m (Either a b)

Lifted race.

Since: 0.1

race_ :: forall m a b. (HasCallStack, MonadAsync m) => m a -> m b -> m ()

Lifted race_.

Since: 0.1

concurrently :: (MonadAsync m, HasCallStack) => m a -> m b -> m (a, b)

Lifted concurrently.

Since: 0.1

concurrently_ :: (MonadAsync m, HasCallStack) => m a -> m b -> m ()

Lifted concurrently_.

Since: 0.1

mapConcurrently

Arguments

:: forall m t a b. (MonadAsync m, Traversable t)

=> (a -> m b)
.

-> t a

-> m (t b)

Lifted mapConcurrently.

Since: 0.1

forConcurrently

Arguments

:: forall m t a b. (MonadAsync m, Traversable t)

=> t a
.

-> (a -> m b)

-> m (t b)

Lifted forConcurrently.

Since: 0.1

mapConcurrently_

Arguments

:: forall m f a b. (MonadAsync m, Foldable f)

=> (a -> m b)
.

-> f a

-> m ()

Lifted mapConcurrently_.

Since: 0.1

forConcurrently_

Arguments

:: forall m f a b. (MonadAsync m, Foldable f)

=> f a
.

-> (a -> m b)

-> m ()

Lifted forConcurrently_.

Since: 0.1

replicateConcurrently :: forall m a. MonadAsync m => Int -> m a -> m [a]

Lifted replicateConcurrently.

Since: 0.1

replicateConcurrently_ :: forall m a. MonadAsync m => Int -> m a -> m ()

Lifted replicateConcurrently_.

Since: 0.1

newtype Concurrently m a

Lifted Concurrently.

Since: 0.1

Constructors

Concurrently

Fields

runConcurrently :: m a

Since: 0.1

compareAsyncs :: Async a -> Async b -> Ordering

Compare two Asyncs that may have different types by their ThreadId.

Specialised operations

STM Operations

data STM a

A monad supporting atomic memory transactions.

waitSTM :: Async a -> STM a

A version of wait that can be used inside an STM transaction.

pollSTM :: Async a -> STM (Maybe (Either SomeException a))

A version of poll that can be used inside an STM transaction.

waitCatchSTM :: Async a -> STM (Either SomeException a)

A version of waitCatch that can be used inside an STM transaction.

Waiting for multiple Asyncs

waitAny :: (MonadAsync m, HasCallStack) => [Async a] -> m (Async a, a)

Lifted waitAny.

Since: 0.1

waitAnyCatch

Arguments

:: (MonadAsync m, HasCallStack)

=> [Async a]
.

-> m (Async a, Either SomeException a)

Lifted waitAnyCatch.

Since: 0.1

waitAnyCancel :: (MonadAsync m, HasCallStack) => [Async a] -> m (Async a, a)

Lifted waitAnyCancel.

Since: 0.1

waitAnyCatchCancel

Arguments

:: (MonadAsync m, HasCallStack)

=> [Async a]
.

-> m (Async a, Either SomeException a)

Lifted waitAnyCatchCancel.

Since: 0.1

waitEither :: (MonadAsync m, HasCallStack) => Async a -> Async b -> m (Either a b)

Lifted waitEither.

Since: 0.1

waitEitherCatch

Arguments

:: (MonadAsync m, HasCallStack)

=> Async a
.

-> Async b

-> m (Either (Either SomeException a) (Either SomeException b))

Lifted waitEitherCatch.

Since: 0.1

waitEitherCancel :: (MonadAsync m, HasCallStack) => Async a -> Async b -> m (Either a b)

Lifted waitEitherCancel.

Since: 0.1

waitEitherCatchCancel

Arguments

:: (MonadAsync m, HasCallStack)

=> Async a
.

-> Async b

-> m (Either (Either SomeException a) (Either SomeException b))

Lifted waitEitherCatchCancel.

Since: 0.1

waitEither_ :: (MonadAsync m, HasCallStack) => Async a -> Async b -> m ()

Lifted waitEither_.

Since: 0.1

waitBoth :: (MonadAsync m, HasCallStack) => Async a -> Async b -> m (a, b)

Lifted waitBoth.

Since: 0.1

Waiting for multiple Asyncs in STM

waitAnySTM :: [Async a] -> STM (Async a, a)

A version of waitAny that can be used inside an STM transaction.

Since: async-2.1.0

waitAnyCatchSTM :: [Async a] -> STM (Async a, Either SomeException a)

A version of waitAnyCatch that can be used inside an STM transaction.

Since: async-2.1.0

waitEitherSTM :: Async a -> Async b -> STM (Either a b)

A version of waitEither that can be used inside an STM transaction.

Since: async-2.1.0

waitEitherCatchSTM :: Async a -> Async b -> STM (Either (Either SomeException a) (Either SomeException b))

A version of waitEitherCatch that can be used inside an STM transaction.

Since: async-2.1.0

waitEitherSTM_ :: Async a -> Async b -> STM ()

A version of waitEither_ that can be used inside an STM transaction.

Since: async-2.1.0

waitBothSTM :: Async a -> Async b -> STM (a, b)

A version of waitBoth that can be used inside an STM transaction.

Since: async-2.1.0

Low-level API

Spawning (low-level API)

async :: (MonadAsync m, HasCallStack) => m a -> m (Async a)

Lifted async.

Since: 0.1

asyncBound :: (MonadAsync m, HasCallStack) => m a -> m (Async a)

Lifted asyncBound.

Since: 0.1

asyncOn :: (MonadAsync m, HasCallStack) => Int -> m a -> m (Async a)

Lifted asyncOn.

Since: 0.1

asyncWithUnmask :: (MonadAsync m, HasCallStack) => ((forall b. m b -> m b) -> m a) -> m (Async a)

Lifted asyncWithUnmask.

Since: 0.1

asyncOnWithUnmask :: (MonadAsync m, HasCallStack) => Int -> ((forall b. m b -> m b) -> m a) -> m (Async a)

Lifted asyncOnWithUnmask.

Since: 0.1

Linking

link :: (MonadAsync m, HasCallStack) => Async a -> m ()

Lifted link.

Since: 0.1

linkOnly :: (MonadAsync m, HasCallStack) => (SomeException -> Bool) -> Async a -> m ()

Lifted linkOnly.

Since: 0.1

link2 :: (MonadAsync m, HasCallStack) => Async a -> Async b -> m ()

Lifted link2.

Since: 0.1

link2Only :: (MonadAsync m, HasCallStack) => (SomeException -> Bool) -> Async a -> Async b -> m ()

Lifted link2Only.

Since: 0.1

data ExceptionInLinkedThread

Constructors

ExceptionInLinkedThread (Async a) SomeException

Pooled concurrency

These functions mirror those defined in unliftio.

pooledMapConcurrentlyN

Arguments

:: (MonadAsync m, HasCallStack, Traversable t)

=> Positive Int
.

-> (a -> m b)

-> t a

-> m (t b)

Lifted pooledMapConcurrentlyN.

Since: 0.1

pooledMapConcurrently

Arguments

:: (MonadAsync m, HasCallStack, Traversable t)

=> (a -> m b)
.

-> t a

-> m (t b)

Lifted pooledMapConcurrently.

Since: 0.1

pooledMapConcurrentlyN_

Arguments

:: (MonadAsync m, Foldable f, HasCallStack)

=> Positive Int
.

-> (a -> m b)

-> f a

-> m ()

Lifted pooledMapConcurrentlyN_.

Since: 0.1

pooledMapConcurrently_

Arguments

:: (MonadAsync m, Foldable f, HasCallStack)

=> (a -> m b)
.

-> f a

-> m ()

Lifted pooledMapConcurrently_.

Since: 0.1

pooledForConcurrentlyN

Arguments

:: forall m t a b. (HasCallStack, MonadAsync m, Traversable t)

=> Positive Int
Max threads > 0

-> t a

-> (a -> m b)

-> m (t b)

Lifted pooledForConcurrentlyN.

Since: 0.1

pooledForConcurrently

Arguments

:: forall m t a b. (HasCallStack, MonadAsync m, Traversable t)

=> t a
.

-> (a -> m b)

-> m (t b)

Lifted pooledForConcurrently.

Since: 0.1

pooledForConcurrentlyN_

Arguments

:: forall m f a b. (Foldable f, HasCallStack, MonadAsync m)

=> Positive Int
Max threads > 0

-> f a

-> (a -> m b)

-> m ()

Lifted pooledForConcurrentlyN_.

Since: 0.1

pooledForConcurrently_

Arguments

:: forall m f a b. (Foldable f, HasCallStack, MonadAsync m)

=> f a
.

-> (a -> m b)

-> m ()

Lifted pooledForConcurrently_.

Since: 0.1

pooledReplicateConcurrentlyN

Arguments

:: forall m a. (HasCallStack, MonadAsync m)

=> Positive Int
Max threads > 0

-> Int
Number of times to perform the action.

-> m a

-> m [a]

Lifted pooledReplicateConcurrentlyN.

Since: 0.1

pooledReplicateConcurrently

Arguments

:: forall m a. (HasCallStack, MonadAsync m)

=> Positive Int
Number of times to perform the action.

-> m a

-> m [a]

Lifted pooledReplicateConcurrently.

Since: 0.1

pooledReplicateConcurrentlyN_

Arguments

:: forall m a. (HasCallStack, MonadAsync m)

=> Positive Int
Max threads > 0

-> Int
Number of times to perform the action.

-> m a

-> m ()

Lifted pooledReplicateConcurrentlyN_.

Since: 0.1

pooledReplicateConcurrently_

Arguments

:: forall m a. (HasCallStack, MonadAsync m)

=> Int
Number of times to perform the action.

-> m a

-> m ()

Lifted pooledReplicateConcurrently_.

Since: 0.1

Reexports

data SomeException

The SomeException type is the root of the exception type hierarchy. When an exception of type e is thrown, behind the scenes it is encapsulated in a SomeException.

class Monad m => MonadThread (m :: Type -> Type)

Represents thread effects.

Since: effects-thread-0.1

Minimal complete definition

threadDelay, throwTo, getNumCapabilities, setNumCapabilities, threadCapability, myThreadId, labelThread, threadLabel

Positive

The Positive type comes from Numeric.Data.Positive from the smart-math package. See that module for more functionality.

data Positive a where

Newtype wrapper that attaches a Positive invariant to some a. Positive is an ASemigroup and MGroup i.e. supports addition, multiplication, and division.

Since: smart-math-0.1

Bundled Patterns

pattern MkPositive :: a -> Positive a
Unidirectional pattern synonym for Positive. This allows us to pattern match on a positive term without exposing the unsafe internal details.

Since: smart-math-0.1

mkPositive :: (Num a, Ord a) => a -> Maybe (Positive a)

Smart constructor for Positive. Returns Nothing if the second parameter is <= 0.

Examples

Expand

>>> mkPositive 7
Just (UnsafePositive 7)

>>> mkPositive 0
Nothing

Since: smart-math-0.1

mkPositiveTH :: (Integral a, Lift a, Show a) => a -> Code Q (Positive a)

Template haskell for creating a Positive at compile-time.

Examples

Expand

>>> $$(mkPositiveTH 1)
UnsafePositive 1

Since: smart-math-0.1

unsafePositive :: (HasCallStack, Num a, Ord a, Show a) => a -> Positive a

Throws an error when given a value <= 0.

WARNING: Partial

Examples

Expand

>>> unsafePositive 7
UnsafePositive 7

Since: smart-math-0.1

(+!) :: (HasCallStack, Num a, Ord a, Show a) => a -> Positive a infixl 7

Postfix operator for unsafePositive.

WARNING: Partial

Examples

Expand

>>> (7 +!)
UnsafePositive 7

Since: smart-math-0.1